Made the navbar a seperate component and made routes direct
